+++ inductiva/resources/__init__.py
@@ -1,3 +1,3 @@
 #pylint: disable=missing-module-docstring
 from .machine_groups import MachineGroup, ElasticMachineGroup, MPICluster, get
-from .utils import estimate_machine_cost, get_available_machine_types, list_available_machines
+from .utils import estimate_machine_cost, get_available_machine_types
