+++ inductiva/resources/utils.py
@@ -39,18 +39,6 @@ class MachineTypeInfo:
         return getattr(self, item)
 
 
-def list_available_machines(provider: Union[str, ProviderType]):
-    """List all available machines types."""
-
-    resources_available = get_available_machine_types(provider)
-    machine_types = []
-
-    for machine in resources_available:
-        machine_types.append(machine["machine_type"])
-
-    return tuple(machine_types)
-
-
 def get_available_machine_types(
     provider: Union[str, ProviderType] = ProviderType.GCP,
     machine_families: Optional[List[str]] = None,
@@ -153,26 +141,3 @@ def estimate_machine_cost(machine_type: str,
 
     return float(estimated_cost)
 
-
-def get_machine_dict(machines):
-    """Get a dictionary with the information of the machines."""
-    column_names = [
-        "Host Name",
-        "Started",
-        "Status",
-        "Last seen",
-        "Running task",
-    ]
-    table = defaultdict(list, {key: [] for key in column_names})
-    for machine in machines:
-        status = "Off" if isinstance(machine["terminated_at"],
-                                     str) else "Active"
-        task_id = machine["current_task_id"] if isinstance(
-            machine["current_task_id"], str) else None
-        table["Host Name"].append(machine["host_name"])
-        table["Started"].append(machine["started_at"])
-        table["Status"].append(status)
-        table["Last seen"].append(machine["last_seen_at"])
-        table["Running task"].append(task_id)
-
-    return table
